@@ -9,10 +9,10 @@
     "cypress": "cypress open",
     "cypress:ci": "run-p --race start:ci cypress:run",
     "cypress:ci-accessibility": "run-p --race start:ci cypress:run-accessibility",
-    "cypress:run": "sleep 10 && cypress run --config '{\"ignoreTestFiles\": \"**/a11y.spec.ts\"}'",
-    "cypress:run-accessibility": "sleep 10 && cypress run --config '{\"testFiles\": \"**/a11y.spec.ts\"}'",
-    "dist": "ng build --build-optimizer=false --progress=false --configuration production && cp nginx.conf mime.types dist",
-    "dist-localised": "ng build --build-optimizer=false --progress=false --configuration production --localize=true && cp nginx.conf mime.types dist",
+    "cypress:run": "sleep 10 && cypress run --config '{\"component\": {},  \"e2e\": { \"excludeSpecPattern\": \"**/**a11y.cy.ts\"}}'",
+    "cypress:run-accessibility": "sleep 10 && cypress run --spec **/*a11y.cy.ts",
+    "dist": "ng build --build-optimizer=false --progress=false --configuration production",
+    "dist-localised": "ng build --build-optimizer=false --progress=false --configuration production --localize=true",
     "i18n:extract": "ng extract-i18n --output-path src/locale --out-file messages.en-GB.xlf",
     "lint": "eslint --ext .js,.ts src/app --fix",
     "lint:ci": "eslint --ext .ts src/app",
@@ -32,44 +32,45 @@
   "dependencies": {
     "@amcharts/amcharts4": "^4.10.22",
     "@amcharts/amcharts4-geodata": "^4.1.21",
-    "@angular/animations": "~12.2.7",
-    "@angular/cdk": "^11.2.13",
-    "@angular/common": "~12.2.7",
-    "@angular/compiler": "^12.2.7",
-    "@angular/core": "~12.2.7",
-    "@angular/forms": "~12.2.7",
-    "@angular/localize": "^12.2.13",
-    "@angular/material": "^11.2.13",
-    "@angular/platform-browser": "~12.2.7",
-    "@angular/platform-browser-dynamic": "~12.2.7",
-    "@angular/router": "~12.2.7",
-    "html2canvas": "^1.3.2",
+    "@angular/animations": "^14.2.0",
+    "@angular/cdk": "^14.2.0",
+    "@angular/common": "^14.2.0",
+    "@angular/compiler": "^14.2.0",
+    "@angular/core": "^14.2.0",
+    "@angular/forms": "^14.2.0",
+    "@angular/localize": "^14.2.0",
+    "@angular/material": "^14.2.0",
+    "@angular/platform-browser": "^14.2.0",
+    "@angular/platform-browser-dynamic": "^14.2.0",
+    "@angular/router": "^14.2.0",
     "pdfmake": "^0.1.72",
     "rxjs": "^6.6.7",
+    "ts-loader": "^9.3.1",
     "tslib": "^2.0.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^12.2.7",
-    "@angular/cli": "^12.2.7",
-    "@angular/compiler-cli": "^12.2.7",
+    "@angular-devkit/build-angular": "^14.2.1",
+    "@angular/cli": "^14.2.1",
+    "@angular/compiler-cli": "^14.2.0",
+    "@cypress/webpack-preprocessor": "^5.12.2",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "^2.0.10",
     "@types/node": "^12.20.26",
-    "@typescript-eslint/eslint-plugin": "^4.31.2",
-    "@typescript-eslint/eslint-plugin-tslint": "^4.31.2",
-    "@typescript-eslint/parser": "^4.31.2",
-    "@typescript-eslint/types": "^4.31.2",
+    "@typescript-eslint/eslint-plugin": "^5.36.0",
+    "@typescript-eslint/eslint-plugin-tslint": "^5.36.0",
+    "@typescript-eslint/parser": "^5.36.0",
+    "@typescript-eslint/types": "^5.36.0",
     "axe-core": "^4.4.1",
     "codelyzer": "^6.0.0",
-    "cypress": "^9.5.3",
+    "cypress": "^10.5.0",
     "cypress-axe": "^0.14.0",
-    "eslint": "^7.32.0",
-    "eslint-plugin-rxjs": "^2.1.7",
+    "eslint": "^8.23.0",
+    "eslint-plugin-rxjs": "^5.0.2",
     "jasmine": "^3.9.0",
     "jasmine-core": "^3.6.0",
     "jasmine-spec-reporter": "~5.0.0",
-    "karma": "^6.3.4",
+    "karma": "^6.4.0",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.0",
@@ -78,8 +79,7 @@
     "npm-run-all": "^4.1.5",
     "prettier": "^2.4.1",
     "ts-node": "~8.3.0",
-    "tslint": "~6.1.0",
-    "typescript": "^4.3.5",
+    "typescript": "^4.7.4",
     "typescript-eslint": "0.0.1-alpha.0"
   }
 }
