@@ -11,19 +11,17 @@ describe('APIService', () => {
   let service: APIService;
   let http: HttpTestingController;
 
-  beforeEach(
-    waitForAsync(() => {
-      TestBed.configureTestingModule({
-        providers: [APIService],
-        imports: [HttpClientTestingModule]
-      }).compileComponents();
-      service = TestBed.inject(APIService);
-      http = TestBed.inject(HttpTestingController);
-    })
-  );
+  beforeEach(waitForAsync(() => {
+    TestBed.configureTestingModule({
+      providers: [APIService],
+      imports: [HttpClientTestingModule]
+    }).compileComponents();
+    service = TestBed.inject(APIService);
+    http = TestBed.inject(HttpTestingController);
+  }));
 
   it('should load the breakdowns', () => {
-    const baseUrl = `${environment.serverAPI}${service.suffixFiltering}`;
+    const baseUrl = `${environment.serverAPI}/${service.suffixFiltering}`;
     const sub = service
       .getBreakdowns({ filters: {} })
       .subscribe((res: BreakdownResults) => {
@@ -35,7 +33,7 @@ describe('APIService', () => {
   });
 
   it('should load the general results', () => {
-    const baseUrl = `${environment.serverAPI}${service.suffixGeneral}`;
+    const baseUrl = `${environment.serverAPI}/${service.suffixGeneral}`;
     const sub = service.getGeneralResults().subscribe((res: GeneralResults) => {
       expect(res).toBeTruthy();
     });
@@ -46,7 +44,7 @@ describe('APIService', () => {
 
   it('should load the general results with content-tier-zero', () => {
     const ctZeroParam = '?content-tier-zero=true';
-    const baseUrl = `${environment.serverAPI}${service.suffixGeneral}${ctZeroParam}`;
+    const baseUrl = `${environment.serverAPI}/${service.suffixGeneral}${ctZeroParam}`;
     const sub = service
       .getGeneralResults(true)
       .subscribe((res: GeneralResults) => {
