@@ -3,8 +3,9 @@
  **/
 
 export const environment = {
+  feedbackUrl: 'https://www.google.com?q=feedbackUrl',
   production: false,
-  serverAPI: 'http://localhost:3001/',
+  serverAPI: 'http://localhost:3001',
   serverPortal: 'https://www.europeana.eu/en/search',
-  wskey: 'api2demo'
+  wskey: 'wskey'
 };
