@@ -1,6 +1,12 @@
+const getEnvVar = (key: string): string | null => {
+  const env = (window as any).__env;
+  return env ? env[key] : null;
+};
+
 export const environment = {
   production: true,
-  serverAPI: 'https://metis-statistics-rest-test.eanadev.org/',
-  serverPortal: 'https://www.europeana.eu/en/search',
-  wskey: 'api2demo'
+  feedbackUrl: getEnvVar('feedbackUrl') || '',
+  serverAPI: getEnvVar('serverAPI') || '',
+  serverPortal: getEnvVar('serverPortal') || '',
+  wskey: getEnvVar('wskey') || ''
 };
