@@ -2,10 +2,15 @@
 /*  settings that can be used throughout the entire application
  **/
 
+const getEnvVar = (key: string): string | null => {
+  const env = (window as any).__env;
+  return env ? env[key] : null;
+};
+
 export const environment = {
   production: false,
-  //serverAPI: 'https://api.europeana.eu/record/v2/search.json',
-  serverAPI: 'https://metis-statistics-rest-test.eanadev.org/',
-  serverPortal: 'https://www.europeana.eu/search',
-  wskey: 'api2demo'
+  feedbackUrl: getEnvVar('feedbackUrl') || '',
+  serverAPI: getEnvVar('serverAPI') || '',
+  serverPortal: getEnvVar('serverPortal') || '',
+  wskey: getEnvVar('wskey') || ''
 };
